exceptions

+++ src/main/java/com/example/kaninosCa/owner/OwnerService.java
@@ -1,12 +1,13 @@
 package com.example.kaninosCa.owner;
 
+import com.example.kaninosCa.exception.BadRequestException;
+import com.example.kaninosCa.exception.NotFoundException;
 import com.example.kaninosCa.pet.Pet;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PostMapping;
 
 import java.util.List;
-import java.util.Optional;
 
 @AllArgsConstructor
 @Service
@@ -20,8 +21,12 @@ public class OwnerService {
 
     @PostMapping
     public void addOwner(Owner owner) {
-        //check if email is taken
-        //throw an error
+        Boolean existsPhone = ownerRepository.selectExistsPhone(owner.getPhone());
+        if (existsPhone){
+            throw new BadRequestException(
+                    "Phone " + owner.getPhone() + " is taken"
+            );
+        }
         ownerRepository.save(owner);
     }
 
@@ -30,6 +35,13 @@ public class OwnerService {
     }
 
     public  void deleteOwner(Long ownerId){
+
+        if (!ownerRepository.existsById(ownerId)) {
+            throw new NotFoundException(
+                    "Owner " + getOwnerById(ownerId).getName() + " does not exists"
+            );
+        }
+
         ownerRepository.deleteById(ownerId);
     }
